requirements.txt
python-dotenv
google-generativeai
elevenlabs
fastapi 
uvicorn
deepgram-sdk==3.*
python-multipart